Many fixes and merge buttons
