+++ src/layout/FileUpload/types.d.ts
@@ -1,6 +1,5 @@
 import type { ExprVal } from 'src/features/expressions/types';
 import type { ComponentTypes, ILayoutCompBase } from 'src/layout/layout';
-import type { IMapping } from 'src/types';
 
 export interface ILayoutCompFileUploadBase<T extends Extract<ComponentTypes, 'FileUpload'>> extends ILayoutCompBase<T> {
   maxFileSizeInMB: number;
@@ -10,8 +9,6 @@ export interface ILayoutCompFileUploadBase<T extends Extract<ComponentTypes, 'Fi
   hasCustomFileEndings?: boolean;
   validFileEndings?: string[] | string;
   alertOnDelete?: ExprVal.Boolean;
-  optionsId?: string;
-  mapping?: IMapping;
 }
 
 export type ILayoutCompFileUpload = ILayoutCompFileUploadBase<'FileUpload'>;
